Code refactoring - adjusted the namespaces to reflect the project folders and removed unused using directives
