+++ src/Plugins/Nop.Plugin.Api/Controllers/CategoriesController.cs
@@ -7,12 +7,12 @@ using System.Web.Http.ModelBinding;
 using Nop.Core.Domain.Catalog;
 using Nop.Plugin.Api.ActionResults;
 using Nop.Plugin.Api.Attributes;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.DTOs.Categories;
 using Nop.Plugin.Api.Helpers;
 using Nop.Plugin.Api.MappingExtensions;
 using Nop.Plugin.Api.ModelBinders;
 using Nop.Plugin.Api.Models.CategoriesParameters;
-using Nop.Plugin.Api.MVC;
 using Nop.Plugin.Api.Serializers;
 using Nop.Plugin.Api.Services;
 using Nop.Services.Catalog;
