+++ src/Plugins/Nop.Plugin.Api/Helpers/IJsonHelper.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace Nop.Plugin.Api.Helpers
+﻿namespace Nop.Plugin.Api.Helpers
 {
     public interface IJsonHelper
     {
