+++ src/Plugins/Nop.Plugin.Api/Helpers/ReflectionHelper.cs
@@ -1,13 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
 using System.Reflection;
-using System.Text;
-using System.Web.Http;
-using System.Web.Http.Results;
-using Nop.Plugin.Api.ContractResolvers;
 
 namespace Nop.Plugin.Api.Helpers
 {
