+++ src/Plugins/Nop.Plugin.Api/Models/CustomersParameters/CustomersParametersModel.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Web.Http.ModelBinding;
 using Newtonsoft.Json;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.ModelBinders;
-using Nop.Plugin.Api.MVC;
 
 namespace Nop.Plugin.Api.Models.CustomersParameters
 {
