+++ src/Plugins/Nop.Plugin.Api/Models/ShoppingCartsParameters/BaseShoppingCartItemsParametersModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Newtonsoft.Json;
-using Nop.Plugin.Api.MVC;
+using Nop.Plugin.Api.Constants;
 
 namespace Nop.Plugin.Api.Models.ShoppingCartsParameters
 {
