+++ src/Plugins/Nop.Plugin.Api/Services/CustomerApiService.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using Nop.Core.Data;
 using Nop.Core.Domain.Customers;
 using Nop.Plugin.Api.DTOs.Customers;
-using Nop.Plugin.Api.MVC;
 using System.Linq;
 using System.Text.RegularExpressions;
 using Nop.Core.Domain.Common;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.DataStructures;
 using Nop.Plugin.Api.Helpers;
 using Nop.Plugin.Api.MappingExtensions;
