+++ src/Plugins/Nop.Plugin.Api/Services/IOrderApiService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Nop.Core.Domain.Orders;
 using Nop.Core.Domain.Payments;
 using Nop.Core.Domain.Shipping;
-using Nop.Plugin.Api.MVC;
+using Nop.Plugin.Api.Constants;
 
 namespace Nop.Plugin.Api.Services
 {
