+++ src/Plugins/Nop.Plugin.Api/Services/IProductCategoryMappingsApiService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using Nop.Core.Domain.Catalog;
-using Nop.Plugin.Api.MVC;
+using Nop.Plugin.Api.Constants;
 
 namespace Nop.Plugin.Api.Services
 {
