+++ src/Plugins/Nop.Plugin.Api/Services/ProductCategoryMappingsApiService.cs
@@ -2,8 +2,8 @@
 using System.Linq;
 using Nop.Core.Data;
 using Nop.Core.Domain.Catalog;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.DataStructures;
-using Nop.Plugin.Api.MVC;
 
 namespace Nop.Plugin.Api.Services
 {
