+++ src/Tests/Nop.Plugin.Api.Tests/ControllersTests/Categories/CategoriesControllerTests_GetCategories.cs
@@ -3,11 +3,11 @@ using System.Web.Http;
 using System.Web.Http.Results;
 using AutoMock;
 using Nop.Core.Domain.Catalog;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.Controllers;
 using Nop.Plugin.Api.DTOs.Categories;
 using Nop.Plugin.Api.MappingExtensions;
 using Nop.Plugin.Api.Models.CategoriesParameters;
-using Nop.Plugin.Api.MVC;
 using Nop.Plugin.Api.Serializers;
 using Nop.Plugin.Api.Services;
 using NUnit.Framework;
