+++ src/Tests/Nop.Plugin.Api.Tests/ControllersTests/Customers/CustomersControllerTests_GetCustomers.cs
@@ -2,10 +2,10 @@
 using System.Collections.Generic;
 using System.Web.Http;
 using System.Web.Http.Results;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.Controllers;
 using Nop.Plugin.Api.DTOs.Customers;
 using Nop.Plugin.Api.Models.CustomersParameters;
-using Nop.Plugin.Api.MVC;
 using Nop.Plugin.Api.Serializers;
 using Nop.Plugin.Api.Services;
 using NUnit.Framework;
