+++ src/Tests/Nop.Plugin.Api.Tests/ControllersTests/Orders/OrdersControllerTests_GetOrders.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Web.Http;
 using System.Web.Http.Results;
 using Nop.Core.Domain.Orders;
+using Nop.Plugin.Api.Constants;
 using Nop.Plugin.Api.Controllers;
 using Nop.Plugin.Api.DTOs.Orders;
 using Nop.Plugin.Api.Models.OrdersParameters;
-using Nop.Plugin.Api.MVC;
 using Nop.Plugin.Api.Serializers;
 using Nop.Plugin.Api.Services;
 using NUnit.Framework;
